+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 This project follows Semantic Versioning.
 
-## [Unreleased]
+## v0.1.0 - 2026-08-27
 
 ### Added